Move topics data file to js/helpers
