+++ client/js/screens/Lessons/SurveyLessons/SurveyLessonsContainer.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { View } from "react-native";
 import SurveyLessons from "./SurveyLessons";
-import topics from "./helper/data";
+import topics from "../../../helpers/topics";
 const SurveyLessonsContainer = () => {
   return (
     <View>
